issue #14 DBへの接続をmodels/connect.goに移動
